# [9.x] Makes Blade inline components 10x faster. ⛽️