--- a/tests/Integration/View/BladeTest.php
+++ b/tests/Integration/View/BladeTest.php
@@ -9,6 +9,13 @@ use Orchestra\Testbench\TestCase;
 
 class BladeTest extends TestCase
 {
+    public function tearDown(): void
+    {
+        Component::flushCache();
+
+        parent::tearDown();
+    }
+
     public function test_rendering_blade_string()
     {
         $this->assertSame('Hello Taylor', Blade::render('Hello {{ $name }}', ['name' => 'Taylor']));