@JayBizzle 2022-10-05
Just tested this against an issue we were having with https://github.com/blade-ui-kit/blade-icons/issues/190

Based on the repo in that issue we are seeing the following improvement when rendering 1000 icons with this PR in place.

```
Before - 280.601ms
After - 173.016ms
```

👍